+++ libs/movie/src/lib/movie/form/media-images/media-images.module.ts
@@ -20,7 +20,6 @@ import { PromotionalLinksModule } from '@blockframes/movie/form/links/promotiona
 // Blockframes Media
 import { FileUploaderModule } from '@blockframes/media/file/file-uploader/file-uploader.module';
 import { ImageUploaderModule } from '@blockframes/media/image/uploader/uploader.module';
-import { ReferencePathModule } from '@blockframes/media/directives/media/reference-path.pipe';
 
 import { MovieFormMediaImagesComponent } from './media-images.component';
 
@@ -36,7 +35,6 @@ import { MovieFormMediaImagesComponent } from './media-images.component';
     FormTableModule,
     ImageUploaderModule,
     FlexLayoutModule,
-    ReferencePathModule,
 
     // Material
     MatIconModule,
