+++ libs/movie/src/lib/movie/marketplace/shell/shell.module.ts
@@ -17,7 +17,7 @@ import { MovieHeaderModule } from '@blockframes/movie/components/header/header.m
 import { PromotionalLinksModule } from '@blockframes/movie/components/promotional-links/promotional-links.module';
 import { WishlistButtonModule } from '@blockframes/organization/components/wishlist-button/wishlist-button.module';
 import { HasKeysModule } from '@blockframes/utils/pipes';
-import { DownloadModule } from '@blockframes/media/pipes/download.pipe';
+import { DownloadPipeModule } from '@blockframes/media/file/pipes/download.pipe';
 import { VideoViewerModule } from '@blockframes/media/video/viewer/viewer.module';
 
 // Material
@@ -43,7 +43,7 @@ import { MatCardModule } from '@angular/material/card';
     CarouselModule,
     MatLayoutModule,
     HasKeysModule,
-    DownloadModule,
+    DownloadPipeModule,
     VideoViewerModule,
 
     // Material
