+++ libs/user/src/lib/auth/forms/profile/profile.module.ts
@@ -6,8 +6,6 @@ import { ReactiveFormsModule } from '@angular/forms';
 
 import { ProfileFormComponent } from './profile.component';
 
-import { ReferencePathModule } from '@blockframes/media/directives/media/reference-path.pipe';
-
 // Material
 import { MatCardModule } from '@angular/material/card';
 import { MatFormFieldModule } from '@angular/material/form-field';
@@ -21,7 +19,7 @@ import { MatInputModule } from '@angular/material/input';
     FlexLayoutModule,
     ImageUploaderModule,
     ReactiveFormsModule,
-    ReferencePathModule,
+
     // Material
     MatCardModule,
     MatFormFieldModule,
